Change max work group size variables type from int to long

+++ tornado-api/src/main/java/uk/ac/manchester/tornado/api/TornadoTargetDevice.java
@@ -55,7 +55,7 @@ public interface TornadoTargetDevice {
 
     long[] getDeviceMaxWorkGroupSize();
 
-    int getMaxThreadsPerBlock();
+    long getMaxThreadsPerBlock();
 
     int getDeviceMaxClockFrequency();
